fix: added config directory in tests and moved common config object there
