+++ packages/accounts/src/kernel-zerodev/__tests__/kernel-kill-switch-provider.test.ts
@@ -6,7 +6,7 @@ import {
   type Hash,
 } from "viem";
 import { polygonMumbai } from "viem/chains";
-import { config } from "./kernel-account.test.js";
+import { config } from "./config/index.js";
 import {
   ECDSAProvider,
   KillSwitchProvider,
