+++ packages/accounts/src/kernel-zerodev/__tests__/kernel-session-key-provider.test.ts
@@ -11,7 +11,7 @@ import {
 } from "viem";
 import { polygonMumbai } from "viem/chains";
 import { generatePrivateKey } from "viem/accounts";
-import { config } from "./kernel-account.test.js";
+import { config } from "./config/index.js";
 import { ECDSAProvider } from "../validator-provider/index.js";
 import {
   CHAIN_ID_TO_NODE,
