+++ packages/accounts/src/kernel-zerodev/__tests__/kernel-validator-provider.test.ts
@@ -1,7 +1,7 @@
 import { createPublicClient, http, type Hex } from "viem";
 import { polygonMumbai } from "viem/chains";
 import { ECDSAValidatorAbi } from "../abis/ESCDAValidatorAbi.js";
-import { config } from "./kernel-account.test.js";
+import { config } from "./config/index.js";
 import { LocalAccountSigner } from "@alchemy/aa-core";
 import { generatePrivateKey } from "viem/accounts";
 import { ECDSAProvider } from "../validator-provider/index.js";
